requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0
django-filter==23.3
dj-rest-auth==5.0.1
django-allauth==0.58.2